someone fix the list feature

+++ app/src/main/java/com/example/fitquest/ui/nutrition/RowEntry.kt
@@ -1,3 +1,6 @@
 package com.example.fitquest.ui.nutrition
 
-data class RowEntry(val textString: String, val textNumber: Float)
+data class RowEntry(
+    val textString: String,
+    val textNumber: Double
+)
